requirements.txt
# Google Cloud Libraries
google-cloud-pubsub==2.29.*
google-api-python-client==2.166.*
google-auth==2.38.*
google-auth-httplib2==0.2.*

python-dotenv==1.1
google-adk==0.1.*

Flask==3.1.*